loader.scheduleLoad and query building
