+++ addon/document/internal/loader.js
@@ -1,7 +1,10 @@
 import Base from './base';
 import ModelMixin from './-model-mixin';
+import OwnerPropertiesMixin from './-owner-properties-mixin';
 
-export default class LoaderInternal extends ModelMixin(Base) {
+// TODO: autoload
+// TODO: load on owner property changes -- observe owner in a mixin from filter
+export default class LoaderInternal extends OwnerPropertiesMixin(ModelMixin(Base)) {
 
   /*
     opts: {
@@ -20,38 +23,46 @@ export default class LoaderInternal extends ModelMixin(Base) {
     this.opts = opts;
   }
 
+  _withPropertyChanges() {
+    return this.parent.withPropertyChanges(...arguments);
+  }
+
+  get _state() {
+    return this.parent.state;
+  }
+
   _createModel() {
     return this.store._createLoader(this);
   }
 
   _willLoad() {
-    this.withPropertyChanges(changed => {
-      this.state.onLoading(changed);
+    this._withPropertyChanges(changed => {
+      this._state.onLoading(changed);
     }, true);
   }
 
-  _didLoad(internal) {
-    this.withPropertyChanges(changed => {
-      this._setContent(internal, changed);
-      this.state.onLoaded(changed);
+  _didLoad() {
+    this._withPropertyChanges(changed => {
+      this._state.onLoaded(changed);
     }, true);
   }
 
   _loadDidFail(err) {
-    this.withPropertyChanges(changed => {
-      this.state.onError(err, changed);
+    this._withPropertyChanges(changed => {
+      this._state.onError(err, changed);
     }, true);
   }
 
   get query() {
-    // TODO: query
+    let properties = this._properties;
+    let query = this.opts.query;
+    return query.call(this.owner, properties);
   }
 
   scheduleLoad() {
     this._willLoad();
-    return this.database._internalDocumentFirst(this.query).then(internal => {
-      // TODO: Don't populate loaded, filter should automatically pick up loaded doc
-      return this._didLoad(internal);
+    return this.database._internalDocumentFirst(this.query).then(() => {
+      return this._didLoad();
     }, err => {
       return this._loadDidFail(err);
     });
